Create XDL Namespace, begin converting writeMeta fn's to constexpr
The xdl namespace has been created for storing anything that is not a
member of XDLType or a subclass of XDLType.

The types array has been converted to a DynArray of const
unique_ptr<XDLType>. The performance of unique_ptr for DynArray seems
negligibly different from using raw pointers after optimizations.
XDLType::classInit has been modified as such to emplace_back a series of
unique_ptrs to the required XDLTypes.

Added a virtual destructor to XDLType. It is the default destructor and
shouldn't do anything, but leaving it out is undefined behavior.

Struct previously took in a XDLCompiler pointer and then did nothing
with the pointer. The pointer is now given to Struct::compiler, which is
the local XDLCompiler pointer.

XDL::XDLDerived is a concept that accepts an XDLType or a subclass of
XDLType. This will be used to simplify several repetitive functions,
such as typeToDataType.

typeToDataType for strings now is generalized to a string_view
regardless of what kind of string it is. This may cause a bit of a
performance hit, but allows us to use string_view::size() instead of
relying on three different flavors of constexpr string length functions
for different types of string.

typeToDataType will soon be generalizable for any XDLDerived, provided
that datatype's global hashmap of XDLType names and types is converted
to a HashMap from an unordered_map. Also, HashMap needs to be made into
a completely constexpr class in order for this to work properly.

GLWin previously called XDLType::classCleanup, but this is no longer
needed as XDLType::classCleanup deleted the raw pointers previously
stored in the types DynArray.