Implement datatype1.hh
XDL types have been autogenerated for a while now, but in the past it has been
split into a separate file for compatibility reasons. datatype.{cc/hh}
has finally been retired and datatype1.{cc,hh} are being implemented on
a trial basis. If everything works out as expected, then the names
will be fixed and this will be the new normal.

datatype1.cc: moves everything into the .hh file. There really is no
reason for this to not be header-only. Also, a header-only
implementation would make it much easier to handle code generation and
much easier to implement constexpr functionality (allowing std's
typeToDataType to work as a constexpr).