Add EnumCounter
EnumCounter: This is a reflection-based method of counting the number of
elements in an enum. This utilizes compile-time template
meta-programming which may increase compile times in some instances. So
far, the compile times seem okay, but with adequate complexity that may
change. This is especially useful for enum classes, because all elements
are contiguous and can therefore be sequentially enumerated. The
alternative for non-enum classes is to check the first 100 elements
(this will eventually be able to be set), and any valid elements are
then counted towards the total.

enumTMP: adds some basic tests to ensure that the counter is working as
expected. Given that EnumCounter is a compile-time object, this should
be changed to use static_asserts instead of printing.

test/CMakeLists.txt: Add enumTMP to list of tests to compile.