Tentative HashMap conversion
HashMap: This has been tentatively converted to use primarily
constexprs functions and has been made header-only. The cc file should
be deleted soon if everything works out as expected. Note: we still
cannot make constexpr HashMaps as "constexpr new" has not been fully
implemented yet. "constexpr new" has been added to GCC 12 and has not
been added to Clang's libcxx as of this commit, which means neither
provides adequate support at the moment. However, this does mean that
HashMap may be able to be used in constexpr functions, as long as we
don't try to define a HashMap itself as constexpr. This will require a
lot of testing before its viable.